refactor experiments
